v1.5.6 -- switch to rmarkdown as the default, use DT package
With rmarkdown v0.9.5 released today and described at
http://blog.rstudio.org/2016/03/21/rmarkdown-v0-9-5/ it doesn't look like
knitrBootstrap adds much anymore as commented by Jim Hester himself at
https://github.com/jimhester/knitrBootstrap/pull/68#issuecomment-199293180

With this in mind, I'm switching the default to rmarkdown instead of using
knitrBootstrap. Specially because the version of knitrBootstrap I like to use
is only available via GitHub and will probably never make it to CRAN.

It might even make sense to completely drop the knitrBootstrap dependency,
although I'll leave it for now since there is no immediate need to drop it.

I'm also testing using the DT package for the datatables features instead of
the JS code that was included before. This might not work on this commit but
I'll get it to work soon anyway.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/regionReport@115061 bc3139a8-67e5-0310-9ffc-ced21a209358